Wrapper enums are default int-based, but Xamarin.Mac is using either long or ulong for their enum base types. Casting through int fixes the issue for these wrappers, which default to int-based enums.

+++ INTV.Shared/View/OSMessageBoxIcon.Mac.cs
@@ -35,27 +35,27 @@ namespace INTV.Shared.View
         /// No icon displayed.
         /// </summary>
         /// <remarks>OK, well, actually, you cannot have 'none' so show info.</remarks>
-        None = NSAlertStyle.Informational,
+        None = (int)NSAlertStyle.Informational,
 
         /// <summary>
         /// Question icon displayed.
         /// </summary>
         /// <remarks>Well... On Mac, it's a warning icon.</remarks>
-        Question = NSAlertStyle.Warning,
+        Question = (int)NSAlertStyle.Warning,
 
         /// <summary>
         /// Information icon displayed.
         /// </summary>
-        Information = NSAlertStyle.Informational,
+        Information = (int)NSAlertStyle.Informational,
 
         /// <summary>
         /// Exclamation icon displayed.
         /// </summary>
-        Exclamation = NSAlertStyle.Warning,
+        Exclamation = (int)NSAlertStyle.Warning,
 
         /// <summary>
         /// Error icon displayed.
         /// </summary>
-        Error = NSAlertStyle.Critical,
+        Error = (int)NSAlertStyle.Critical,
     }
 }
